Moved pipe server and splited it to server and host aka shared between server and client
