+++ src/Common/pipeclient.cpp
@@ -1,0 +1,6 @@
+#include "pipeclient.h"
+
+PipeClient::PipeClient()
+{
+
+}
