+++ src/Common/pipeclient.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "../UI/loginfo.h"
+
+class PipeClient
+{
+public:
+    PipeClient();
+    char ReceivePermission(const DWORD Pid);
+    bool SendLog(const LogInfo& logInfo);
+};
